Troubleshooted Audio Issue
Found issue preventing volume settings from correctly loading.
Each audio setting creates its own AudioSaveGame such that when they save it overrides the existing one causing only the last audio to be saved (music).
Hence need to save game to be in the audio settings class and for saving to be sequential.