113 separate tag ability (#114)
* Added Apply Tag Class

New apply tag gameplay tag added as well.

* Implemented ApplyTagAbility

Implemented the apply tag ability which sends the corresponding gameplay events the the tagged and tagging player respectively.

* Reorganised abilities folder

Sorted abilities into folders.

* Implemented OnEventAbility

This ability activates when it receives the specified event and simply adds and removes specified effects.

* Sent Tag Event

TagAbility sends tag notiofy to applytag ability.
Issue where it doe not work on the client.

* Bypassed effects on tag ability

Effects are not granted or removed in the tag ability, need to fix OnEventAbility not firing.

* Fixed OnEventAbility not firing

Had to set trigger in BP. Had to add correct tag to TaggedEffect. Issue where it does not run on the client.

* Fixed ApplyTag Not Activating on other char

* Fixed OnEventTag

OnEventTag now correctly runs on server and client.